+++ src/main/java/kitchenpos/application/dto/request/TableGroupRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto.request;
+package kitchenpos.application.dto.request;
 
 import java.util.List;
 import java.util.stream.Collectors;
