+++ src/test/java/kitchenpos/application/TableServiceTest.java
@@ -20,10 +20,10 @@ import kitchenpos.domain.OrderLineItem;
 import kitchenpos.domain.OrderStatus;
 import kitchenpos.domain.OrderTable;
 import kitchenpos.domain.TableGroup;
-import kitchenpos.dto.request.ChangeNumOfTableGuestsRequest;
-import kitchenpos.dto.request.ChangeOrderTableEmptyRequest;
-import kitchenpos.dto.request.OrderTableRequest;
-import kitchenpos.dto.response.OrderTableResponse;
+import kitchenpos.application.dto.request.ChangeNumOfTableGuestsRequest;
+import kitchenpos.application.dto.request.ChangeOrderTableEmptyRequest;
+import kitchenpos.application.dto.request.OrderTableRequest;
+import kitchenpos.application.dto.response.OrderTableResponse;
 import kitchenpos.support.SpringBootNestedTest;
 
 @Transactional
